requirements.txt
torchdiffeq>=0.2.3
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0
pyyaml>=6.0
tabulate>=0.9.0
scipy>=1.10.0
lightgbm>=3.3.0  # For ML benchmarks (optional)
tqdm>=4.60.0  # For progress bars
